Successfully implemented user input.
I was banging my head against a wall for a quite some time before
realising that the issue a simple one, I variable used in the
'generateGrid()' function was different to the one used in the for loop
(numberOfSquares & numOfSquares).

Nice! Feeling good.

I've also added a function to generate randomRBG codes for the rainbow
effect. Just need to link this up to the button and do some final
polishing and then the app will be complete.

 # Please enter the commit message for
your changes. Lines starting